Extract common functions and constant from place order CLI
