+++ stf/src/commands/mod.rs
@@ -16,3 +16,6 @@
 */
 
 pub mod place_order;
+
+mod common_args;
+mod common_args_processing;
